# backend/docs/openapi.yaml
openapi: 3.0.3
info:
  title: Places API
  version: 1.0.0
  description: API for managing users and places.
servers:
  - url: http://localhost:5001/api/v1
    description: Local development server
paths:
  /users:
    get:
      operationId: listUsers
      summary: List users
      parameters:
        - $ref: "#/components/parameters/CorrelationId"
      responses:
        "200":
          description: Users returned successfully.
          headers:
            x-correlation-id:
              $ref: "#/components/headers/CorrelationId"
          content:
            application/json:
              schema:
                type: object
                required: [users]
                properties:
                  users:
                    type: array
                    items:
                      $ref: "#/components/schemas/User"
        "500":
          $ref: "#/components/responses/InternalServerError"
    post:
      operationId: createUser
      summary: Create a user
      parameters:
        - $ref: "#/components/parameters/CorrelationId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateUserRequest"
      responses:
        "201":
          description: User created successfully.
          headers:
            x-correlation-id:
              $ref: "#/components/headers/CorrelationId"
          content:
            application/json:
              schema:
                type: object
                required: [user]
                properties:
                  user:
                    $ref: "#/components/schemas/User"
        "422":
          $ref: "#/components/responses/ValidationError"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /places:
    get:
      operationId: listPlaces
      summary: List places
      parameters:
        - name: creatorId
          in: query
          required: false
          schema:
            type: string
          description: Return only places created by this user.
        - $ref: "#/components/parameters/CorrelationId"
      responses:
        "200":
          description: Places returned successfully.
          headers:
            x-correlation-id:
              $ref: "#/components/headers/CorrelationId"
          content:
            application/json:
              schema:
                type: object
                required: [places]
                properties:
                  places:
                    type: array
                    items:
                      $ref: "#/components/schemas/Place"
        "500":
          $ref: "#/components/responses/InternalServerError"
    post:
      operationId: createPlace
      summary: Create a place
      parameters:
        - $ref: "#/components/parameters/CorrelationId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreatePlaceRequest"
      responses:
        "201":
          description: Place created successfully.
          headers:
            x-correlation-id:
              $ref: "#/components/headers/CorrelationId"
          content:
            application/json:
              schema:
                type: object
                required: [place]
                properties:
                  place:
                    $ref: "#/components/schemas/Place"
        "422":
          $ref: "#/components/responses/ValidationError"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /sessions:
    post:
      operationId: loginUser
      summary: Authenticate a user
      parameters:
        - $ref: "#/components/parameters/CorrelationId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: User authenticated successfully.
          headers:
            x-correlation-id:
              $ref: "#/components/headers/CorrelationId"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LoginResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /places/{placeId}:
    parameters:
      - $ref: "#/components/parameters/PlaceId"
    get:
      operationId: getPlace
      summary: Get a place
      parameters:
        - $ref: "#/components/parameters/CorrelationId"
      responses:
        "200":
          description: Place returned successfully.
          headers:
            x-correlation-id:
              $ref: "#/components/headers/CorrelationId"
          content:
            application/json:
              schema:
                type: object
                required: [place]
                properties:
                  place:
                    $ref: "#/components/schemas/Place"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalServerError"
    patch:
      operationId: updatePlace
      summary: Update a place
      parameters:
        - $ref: "#/components/parameters/CorrelationId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdatePlaceRequest"
      responses:
        "200":
          description: Place updated successfully.
          headers:
            x-correlation-id:
              $ref: "#/components/headers/CorrelationId"
          content:
            application/json:
              schema:
                type: object
                required: [place]
                properties:
                  place:
                    $ref: "#/components/schemas/Place"
        "404":
          $ref: "#/components/responses/NotFound"
        "422":
          $ref: "#/components/responses/ValidationError"
        "500":
          $ref: "#/components/responses/InternalServerError"
    delete:
      operationId: deletePlace
      summary: Delete a place
      parameters:
        - $ref: "#/components/parameters/CorrelationId"
      responses:
        "200":
          description: Place deleted successfully.
          headers:
            x-correlation-id:
              $ref: "#/components/headers/CorrelationId"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalServerError"
components:
  parameters:
    CorrelationId:
      name: x-correlation-id
      in: header
      required: false
      schema:
        type: string
        minLength: 1
    PlaceId:
      name: placeId
      in: path
      required: true
      schema:
        type: string
        minLength: 1
  headers:
    CorrelationId:
      description: Identifier used to correlate the request with server logs.
      schema:
        type: string
  responses:
    NotFound:
      description: The requested resource was not found.
      headers:
        x-correlation-id:
          $ref: "#/components/headers/CorrelationId"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    ValidationError:
      description: Request validation failed.
      headers:
        x-correlation-id:
          $ref: "#/components/headers/CorrelationId"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Unauthorized:
      description: Authentication failed.
      headers:
        x-correlation-id:
          $ref: "#/components/headers/CorrelationId"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    InternalServerError:
      description: An unexpected server error occurred.
      headers:
        x-correlation-id:
          $ref: "#/components/headers/CorrelationId"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
  schemas:
    User:
      type: object
      required: [id, name, email]
      properties:
        id:
          type: string
        name:
          type: string
        email:
          type: string
          format: email
    CreateUserRequest:
      type: object
      required: [name, email, password]
      properties:
        name:
          type: string
          minLength: 1
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 6
          writeOnly: true
    LoginRequest:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          writeOnly: true
    Coordinates:
      type: object
      required: [lat, lng]
      properties:
        lat:
          type: number
        lng:
          type: number
    Place:
      type: object
      required: [id, title, description, location, creator]
      properties:
        id:
          type: string
        title:
          type: string
        description:
          type: string
        address:
          type: string
        location:
          $ref: "#/components/schemas/Coordinates"
        creator:
          type: string
    CreatePlaceRequest:
      type: object
      required: [title, description, address]
      properties:
        title:
          type: string
          minLength: 1
        description:
          type: string
          minLength: 5
        address:
          type: string
          minLength: 1
        creator:
          type: string
    UpdatePlaceRequest:
      type: object
      required: [title, description]
      properties:
        title:
          type: string
          minLength: 1
        description:
          type: string
          minLength: 5
    ErrorResponse:
      type: object
      required: [message]
      properties:
        message:
          type: string
    MessageResponse:
      type: object
      required: [message]
      properties:
        message:
          type: string
    LoginResponse:
      type: object
      required: [message]
      properties:
        message:
          type: string